requirements.txt
Pillow==9.5.0
PyMuPDF==1.22.5
streamlit==1.24.0
streamlit-drawable-canvas==0.9.3
watchdog==3.0.0
